feat(type-designer): add card menu component + delete type operation
